pyproject.toml
[tool.poetry]
name = "erasmus"
version = "22.10.1.dev0"
description = ""
authors = ["Bryan Forbes <bryan@reigndropsfall.net>"]
license = "BSD-3-Clause"
packages = [
  { include = "erasmus" }
]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.8"
attrs = "^22.1.0"
beautifulsoup4 = "^4.11"
"botus_receptus" = {git = "https://github.com/bryanforbes/botus_receptus.git", rev = "29b03877068ed401fbde6870270ba369591a4a30", extras = ["uvloop", "sqlalchemy"]}
"discord.py" = { version="^2.0.1", extras = ["speed"] }
more-itertools = "~8.14.0"
"fluent.runtime" = "^0.3.1"

[tool.poetry.group.dev.dependencies]
asyncpg-stubs = "~0.26.4"
sqlalchemy2-stubs = "^0.0.2a29"
black = "^22.10"
coverage = {version = "^6.5", extras = ["toml"]}
flake8 = "~5.0"
flake8-bugbear = ">=21.3.1"
flake8-comprehensions = "^3.10.0"
flake8-implicit-str-concat = "^0.3.0"
flake8-pie = "^0.16.0"
flake8-pyi = "^22.10"
flake8-pytest-style = "^1.6.0"
flake8-simplify = "^0.19.3"
flake8-type-checking = "^2.1.3"
isort = "^5.10"
mypy = ">=0.982"
pre-commit = "^2.20"
pytest = "^7.1"
pytest-asyncio = "^0.19.0"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.0"
pytest-mock = "^3.10"
pytest-recording = "^0.12.1"
pytest-sugar = "^0.9.5"
pytest-xdist = "^2.5"

[tool.poetry.scripts]
erasmus = 'erasmus.run:main'

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | htmlcov
)/
'''

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "^\\s+pass$",
  "^\\s*if TYPE_CHECKING:",
  "^\\s*\\.\\.\\.\\s*$"
]

[tool.isort]
extra_standard_library = ["typing_extensions", "_typeshed"]
known_first_party = ["erasmus"]
known_third_party = ["alembic"]
virtual_env = ".venv"
profile = "black"
combine_as_imports = true
add_imports = ["from __future__ import annotations"]
append_only = true

[tool.pyright]
include = ["erasmus", "tests", "stubs"]
stubPath = "./stubs"

pythonVersion = "3.10"
reportUnnecessaryTypeIgnoreComment = "warning"
typeCheckingMode = "strict"
reportPrivateUsage = "none"
reportUnusedImport = "none"
reportUnusedVariable = "none"
reportImportCycles = "none"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=erasmus --cov-branch --cov-report=term -nauto --dist=loadscope"
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
